requirements.txt
json5==0.9.10
beautifulsoup4--4.11.1
lxml==4.9.0
Pillow==9.0.0